pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "ai-startup-ceo-simulator"
version = "3.0.0"
description = "OpenEnv RL environment — AI agent acts as a startup CEO making strategic decisions"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.29.0",
    "pydantic>=2.0",
    "requests==2.31.0",
    "python-dotenv==1.0.0",
    "websockets>=15.0.1",
    "aiofiles==23.2.1",
    "openenv-core>=0.2.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["env*", "agent*", "grader*", "tasks*"]

[openenv]
name = "ai-startup-ceo-simulator"
version = "3.0.0"
spec_version = "openenv-0.1"
entry_point = "env.environment:app"
inference_entry = "inference.py"
